LUT: continued work on the inclusion of entropy in the look-up-table.
Note that this is still (subtly) borken for low-temperature air, at least.

Also, note that the extrapolation limits for the LUT were removed in this
revision.  The limits were interfering with the estimates of entropy at
room-tempeprature conditions when the table was generated with a lower
temperature limit of 500K, as is the case for the air-ions table.
Be aware that this removal of extrapolation limits may change the calculations
other people have been doing...

+++ lib/cfpylib/gasdyn/test_libgas_gas.py
@@ -42,26 +42,55 @@ class TestLibGasGas(unittest.TestCase):
         testGas = make_gas_from_name('air-thermally-perfect')
         p = 100.0e3 # Pa
         T = 300.0 # degrees K
-        gam = 1.4
-        R = 287 # J/degK.kg
         testGas.set_pT(p, T)
-        self.assertAlmostEqual(testGas.rho, p/(R*T), delta=0.01)
+        gam = testGas.gam
+        R = testGas.R
+        Cp = testGas.C_p
+        self.assertAlmostEqual(gam, 1.4, delta=0.01)
+        self.assertAlmostEqual(R, 287.0, delta=2)
+        rho = testGas.rho
+        self.assertAlmostEqual(rho, p/(R*T), delta=0.01)
         import math
         self.assertAlmostEqual(testGas.a, math.sqrt(gam*R*T), delta=2.0)
+        s = testGas.s
+        p2 = 150.0e3
+        T2 = 350.0
+        testGas.set_pT(p2, T2)
+        s2 = testGas.s
+        delta_s = Cp*math.log(T2/T) - R*math.log(p2/p)
+        self.assertAlmostEqual(s2-s, delta_s, delta=0.3)
+        testGas.set_ps(p2, s)
+        rho2 = testGas.rho
+        self.assertAlmostEqual(p2/pow(rho2,gam), p/pow(rho,gam), delta=100.0)
 
     def test_cea_lut_air(self):
         # Peter J.
-        os.system('cp ~/cfcfd3/lib/gas/cea-cases/cea-lut-air-ions.lua.gz .')
+        if not os.path.exists('cea-lut-air-ions.lua.gz'):
+            os.system('cp ~/cfcfd3/lib/gas/cea-cases/cea-lut-air-ions.lua.gz .')
         testGas = Gas('cea-lut-air-ions.lua.gz')
         p = 100.0e3 # Pa
         T = 300.0 # degrees K
-        gam = 1.4
-        R = 287 # J/degK.kg
         testGas.set_pT(p, T)
-        self.assertAlmostEqual(testGas.rho, p/(R*T), delta=0.01)
+        gam = testGas.gam
+        R = testGas.R
+        Cp = testGas.C_p
+        self.assertAlmostEqual(gam, 1.4, delta=0.02)
+        self.assertAlmostEqual(R, 287.0, delta=2)
+        rho = testGas.rho
+        self.assertAlmostEqual(rho, p/(R*T), delta=0.01)
         import math
         self.assertAlmostEqual(testGas.a, math.sqrt(gam*R*T), delta=2.0)
-        
+        s = testGas.s
+        p2 = 150.0e3
+        T2 = 350.0
+        testGas.set_pT(p2, T2)
+        s2 = testGas.s
+        delta_s = Cp*math.log(T2/T) - R*math.log(p2/p)
+        self.assertAlmostEqual(s2-s, delta_s, delta=0.3)
+        testGas.set_ps(p2, s)
+        rho2 = testGas.rho
+        self.assertAlmostEqual(p2/pow(rho2,gam), p/pow(rho,gam), delta=100.0)
+       
 if __name__ == '__main__':
     suite = unittest.TestLoader().loadTestsFromTestCase(TestLibGasGas)
     unittest.TextTestRunner(verbosity=2).run(suite)
